Add  NoDisplay=true
quote from https://specifications.freedesktop.org/desktop-entry-spec/latest/ar01s06.html

NoDisplay means "this application exists, but don't display it in the menus". This can be useful to e.g. associate this application with MIME types, so that it gets launched from a file manager (or other apps), without having a menu entry for it (there are tons of good reasons for this, including e.g. the netscape -remote, or kfmclient openURL kind of stuff).